requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
firebase-admin==6.1.0
Flask==2.3.1
frozenlist==1.3.3
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.2
msgpack==1.0.5
multidict==6.0.4
oauth2client==3.0.0
oauthlib==3.2.2
openai==0.27.5
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase==3.0.27
python-dotenv==1.0.0
python-jwt==2.0.1
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.2
yarl==1.9.2
zipp==3.15.0
